Remove slider focus actions to avoid click bug
Work around for bug in Chrome where clicking slider would not position
it. This fixes the issue by removing focus/blur events and preventing
control from being rerendered. In future, we should try to fix this the
right way by allowing focus/blur to be represented in model and
rendering it with an effect.