Fix batch of actions for starting a recipe on a fresh state
If there is no active recipe when a new one is started, the UI briefly shows an
error, as well as an error in the JS console, as calling the API endpoint to
perform ROS service (/stop_recipe) calls fails.

Now if is no active recipe, just start the new one, otherwise stop the active
one and then start the new one.