Use dismissable banner for connection problems
We'll avoid use of refreshable banners until we have a way to clear
banners by a particular id/role. We don't want to clear the banner when
we get back a good connection because we don't want to accidentally
clear some other banner that was thrown up for some other reason.